Update App.js
A significant portion of the profits from transaction fees will be donated to good causes in Israel that focus on areas like nature preservation and the restoration of old cities, rather than funding the war. We are committed to ensuring that our contributions do not sponsor the war, focusing instead on positive, non-military initiatives that benefit Israel’s heritage and environment.